+++ web/modules/custom/ai_screening_fixtures_base/src/Fixture/TermDepartmentFixture.php
@@ -25,4 +25,11 @@ class TermDepartmentFixture extends TaxonomyTermFixture {
     'Department F',
   ];
 
+  /**
+   * {@inheritdoc}
+   */
+  public function getGroups() {
+    return ['taxonomy_department'];
+  }
+
 }