+++ web/themes/custom/itkdev/itkdev_project_theme/js/project-scripts.js
@@ -3,3 +3,4 @@ require("./selectColorBlock");
 require("./disableAutoEvaluation.js");
 require("./choices.js");
 require("./charts.js");
+require("./headerIcons.js");