+++ web/themes/custom/itkdev/itkdev_project_theme/tailwind.config.js
@@ -13,6 +13,7 @@ module.exports = {
     "bg-red-700",
     "bg-green-600",
     "bg-green-700",
+    "bg-zinc-300",
     "grid-cols-4",
     "order-last",
     "text-slate-300",
@@ -21,7 +22,7 @@ module.exports = {
     "border-yellow-600",
     "green-600",
     "green-700",
-    "border-green-600",
+    "border-green-600"
   ],
   theme: {
     colors: {