+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+- [PR-88](https://github.com/itk-dev/ai-screening/pull/88)
+  - Update report for law track
+  - Update maa_vi webform
 - [PR-87](https://github.com/itk-dev/ai-screening/pull/87)
   - Security updates
 - [PR-85](https://github.com/itk-dev/ai-screening/pull/85)